extract serialization tests for CompositeFilterDescriptor
